+++ examples/photos.js
@@ -17,6 +17,9 @@ const createPhotoPiles = async element => {
 
   piling.set('pileBorderSize', pile => pile.items.length - 1);
 
+  piling.set('pileItemAlignment', false);
+  piling.set('pileItemRotation', true);
+
   return piling;
 };