+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### Next
 
+- Add `randomOffsetRange` and `randomRotationRange` properties
+- Animate item rotation
 - Fix pile scaling
 
 ## v0.5.0
@@ -10,6 +12,7 @@
   - `pileBorderSize`
   - `itemOpacity`
 - Rename the following properties for clarity:
+
   - `itemPadding` is now called `cellPadding`
   - `pileCellAlign` is now called `pileCellAlignment`
   - `itemAlignment` is now called `pileItemAlignment`